// src/app/features/mastertables/tblHSNMaster/models/tblHSNMaster.model.ts
export interface TblHSNMaster {
	fldId: number;
	fldFKHSNCategory: number;
	fldItemNames: string;
	fldSalesCode: string;
	fldJobworkCode: string;
	fldGSTSalesPercentage: number;
	fldGSTJobworkPercentage: number;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
	fldDeletedBy: number;
	fldDeletedDt: Date;
}